# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: session-broker
  version: 0.1.0
  summary: Brokers browser sessions against an upstream OIDC provider and takes custody of the upstream tokens.
  description: |
    Three properties define the service, in priority order:

    1. **Non-invalidating rotation.** Issuing a new session cookie does not
       invalidate the previous one; generations overlap for a grace window.
       This removes the refresh race class outright rather than coordinating
       around it.
    2. **Local-only fast refresh.** Session lifetime is decoupled from upstream
       token lifetime. `POST /session/refresh` performs no network I/O (INV-8).
    3. **Server-side guarantees, dumb client.** The browser SDK needs no
       cross-tab coordination beyond one Web Lock.

    ## Cookies

    - `__Host-broker_session` (`broker_session` over loopback) — `HttpOnly`,
      `SameSite=Lax`. The credential. Never readable by script.
    - `broker_meta` — deliberately readable by script (INV-9), and deliberately
      **unsigned**. It is a client hint. The server never reads it back, and
      signing it would imply it could be trusted as input.

    ## CSRF

    Same-origin path-mounted deployment is the default (INV-10, ADR-0012), so
    there is no CORS surface. The CSRF boundary is `Sec-Fetch-Site` / `Origin`
    on unsafe methods (ADR-0004), which the browser sets and script cannot forge.
  license:
    name: MIT OR Apache-2.0
servers:
  - url: https://app.example.com
...
paths:
...
  /session:
    get:
      tags: [session]
      summary: The current session's meta.
      description: Read-only. Answers from memory; performs no I/O.
      responses:
        "200":
          description: A live session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionMeta" }
        "401":
          description: No session, or one that no longer authenticates.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /session/refresh:
    post:
      tags: [session]
      summary: Rotate (or coalesce onto) the current generation.
      description: |
        **Performs no network I/O** (INV-8) — it never calls the IdP. Refreshes
        arriving within the coalesce window re-issue the current generation
        rather than minting, so N concurrent refreshes mint exactly once.

        Rotation is non-invalidating: the previous cookie stays valid for the
        grace window, which is why a client needs no cross-tab coordination.
      responses:
        "200":
          description: Session refreshed. Both cookies re-issued.
          headers:
            Set-Cookie: { schema: { type: string } }
            X-Broker-Handler-Us:
              description: Server-side handler time in microseconds, for the demo's RTT panel.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionMeta" }
        "401":
          description: Unknown token, a retired generation, or an expired session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: The CSRF boundary refused this request (ADR-0004).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    get:
      tags: [session]
      summary: Interactive refresh — a browser navigation, not a fetch.
      parameters:
        - { name: interactive, in: query, schema: { type: string, enum: ["1"] } }
        - { name: return_to, in: query, schema: { type: string } }
      responses:
        "303":
          description: Redirect back to `return_to`, or to login if the session is gone.
...
  /admin/sessions:
    get:
      tags: [admin]
      summary: List live sessions.
      description: |
        Read from the store, not the in-memory map: the map holds token hashes,
        and an admin surface has no business near credential material even in
        hashed form.
      security: [{ adminApiKey: [] }]
      responses:
        "200":
          description: Live sessions.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminSession" }
        "401": { $ref: "#/components/responses/AdminUnauthorised" }
  /admin/sessions/{sid}:
    delete:
      tags: [admin]
      summary: Force-logout a session.
      description: Kills every generation at once (INV-7), in memory and durably.
      security: [{ adminApiKey: [] }]
      parameters:
        - { name: sid, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Session revoked. }
        "401": { $ref: "#/components/responses/AdminUnauthorised" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-broker_session
      description: "`HttpOnly`. Sent automatically on same-origin requests; script can never read it."
    brokerApiKey:
      type: http
      scheme: bearer
      description: |
        A backend API key — issued via `/admin/api-keys`, or the bootstrap key
        from config. Required IN ADDITION to a session token; neither alone is
        sufficient.
    adminApiKey:
      type: http
      scheme: bearer
      description: |
        The admin key. Deliberately a DIFFERENT value from the backend key: one
        that can mint keys is strictly more powerful than one that exchanges a
        session for a token, and sharing them would let every backend service
        issue credentials for itself.
  responses:
    AdminUnauthorised:
      description: "`unauthorised` — a valid admin API key is required."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/LaneError" }
...
  schemas:
    SessionMeta:
      type: object
      description: |
        Also the decoded content of the `broker_meta` cookie. Non-secret by
        construction and unsigned on purpose (INV-9) — the server never reads it
        back, so nothing here is a basis for a decision.
      required: [v, sub, sid, gen, active_until, refresh_until, absolute_until, custody]
      properties:
        v: { type: integer, const: 1 }
        sub: { type: string }
        sid: { type: string }
        gen: { type: integer, description: Generation number; increments on each rotation. }
        active_until: { type: integer, description: Unix seconds — when this generation stops authenticating. }
        refresh_until: { type: integer, description: Idle expiry; slides on use. }
        absolute_until: { type: integer, description: Hard ceiling; never slides. }
        custody:
          type: string
          enum: [ok, degraded, dead]
          description: Health of the upstream grant behind this session.
...
    AdminSession:
      type: object
      required: [sid, sub, custody_id, current_gen, idle_exp, absolute_exp]
      properties:
        sid: { type: string }
        sub: { type: string }
        custody_id: { type: string }
        current_gen: { type: integer }
        idle_exp: { type: integer }
        absolute_exp: { type: integer }
...
    ApiError:
      type: object
      description: The public lane's error shape.
      properties:
        code: { type: string }
        message: { type: string }
        login_url:
          type: string
          nullable: true
          description: Where to send the browser to re-authenticate, so the SDK never builds one itself.
    LaneError:
      type: object
      description: The internal and admin lanes' error shape.
      required: [error, detail]
      properties:
        error:
          type: string
          description: A stable machine-readable code. Clients should branch on this, not on the status alone.
        detail: { type: string }
security:
  - sessionCookie: []